requirements.txt
alembic==1.8.1
certifi==2022.9.24
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
clip @ git+https://github.com/openai/CLIP.git@d50d76daa670286dd6cacf3bcd80b5e4823fc8e1
Flask==2.2.2
Flask-Login==0.6.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==3.0.0
ftfy==6.1.1
googletrans==4.0.0rc1
greenlet==1.1.3
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.10.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.3
MarkupSafe==2.1.1
numpy==1.23.3
pandas==1.5.0
Pillow==9.2.0
psycopg2==2.9.4
python-dateutil==2.8.2
pytz==2022.4
regex==2022.9.13
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
torch==1.13.0
torchvision==0.14.0
tqdm==4.64.1
typing_extensions==4.3.0
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
zipp==3.8.1
